requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.1
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
numpy==1.21.6
openpyxl==3.1.2
pandas==1.3.5
paramiko==3.1.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0